+++ abramov.vladislav/S7/edge.hpp
@@ -15,16 +15,4 @@ namespace abramov
     Edge(const std::string &f, const std::string &t, size_t w);
   };
 }
-
-abramov::Edge::Edge():
-  from(std::string{}),
-  to(std::string{}),
-  weight(0)
-{}
-
-abramov::Edge::Edge(const std::string &f, const std::string &t, size_t w):
-  from(f),
-  to(t),
-  weight(w)
-{}
 #endif
